#3 poc, simplify type bounds

+++ src/main/scala/essent/credit/package.scala
@@ -3,7 +3,7 @@ package essent
 package object credit {
 
   // Common
-  type Amount            = BigDecimal            // e.g. 10.00
+  type Amount            = Long                  // e.g. 10.00 euro is 1000
   type IBAN              = String                // e.g. NL19ABNA0573066809
   type UUID              = java.util.UUID        // e.g. 47b16912-eb14-483c-86fb-4c9e4de31fbd
   type Date              = String                // e.g. 2016-07-24, i.e. `YYYY-MM-DD` representation
